program: sequence check as u8 (#909)
* program: sequence check as u8 (reuse existing padding)

(cherry picked from commit 0728bb566f6e98d89d730ce64a5dd25f706135e9)

+++ programs/mango-v4/src/instructions/sequence_check.rs
@@ -4,7 +4,7 @@ use crate::accounts_ix::*;
 use crate::error::MangoError;
 use crate::state::*;
 
-pub fn sequence_check(ctx: Context<SequenceCheck>, expected_sequence_number: u64) -> Result<()> {
+pub fn sequence_check(ctx: Context<SequenceCheck>, expected_sequence_number: u8) -> Result<()> {
     let mut account = ctx.accounts.account.load_full_mut()?;
 
     require_eq!(
